--- teardrop/takecare.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.6.33829.357
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "takecare", "takecare\takecare.csproj", "{82E9CFB5-D664-4C3E-9B55-3B7FBA97EFC2}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "teardrop", "teardrop\teardrop.csproj", "{53CB1C2B-CD49-4364-8E57-D8DD3C2553D3}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{82E9CFB5-D664-4C3E-9B55-3B7FBA97EFC2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{82E9CFB5-D664-4C3E-9B55-3B7FBA97EFC2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{82E9CFB5-D664-4C3E-9B55-3B7FBA97EFC2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{82E9CFB5-D664-4C3E-9B55-3B7FBA97EFC2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{53CB1C2B-CD49-4364-8E57-D8DD3C2553D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{53CB1C2B-CD49-4364-8E57-D8DD3C2553D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{53CB1C2B-CD49-4364-8E57-D8DD3C2553D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{53CB1C2B-CD49-4364-8E57-D8DD3C2553D3}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {70631BDE-23B0-448E-AB3B-985AC46884F6}
-	EndGlobalSection
-EndGlobal
